requirements.txt
flask
scikit-learn
pandas
numpy
nltk
PyPDF2
python-docx
joblib
werkzeug
